change all necessary int to long long int

+++ cpp/NSKalgo.cpp
@@ -5,11 +5,11 @@
 #include <cstdlib> // for rand() and srand()
 
 // Primes used in key generation
-static const std::vector<int> pval = {2, 3, 5, 7, 11, 13, 17, 19};
+static const std::vector<long long int> pval = {2, 3, 5, 7, 11, 13, 17, 19};
 
 // Secret key generation (finds secret key 's' such that GCD(s, p-1) = 1)
-int NSKalgo::secretkey(int p) {
-    int s = std::rand() % (p - 1);  // Random between 0 and p-1
+long long int NSKalgo::secretkey(long long int p) {
+    long long int s = std::rand() % (p - 1);  // Random between 0 and p-1
     while (MathFunc::itemgcd(p - 1, s) != 1) {
         s = (s + 1) % p;
     }
@@ -17,9 +17,9 @@ int NSKalgo::secretkey(int p) {
 }
 
 // Public key generation using modular exponentiation
-std::vector<int> NSKalgo::publickey(int p, int s) {
-    std::vector<int> temp(pval.size());
-    int x, y;
+std::vector<long long int> NSKalgo::publickey(long long int p, long long int s) {
+    std::vector<long long int> temp(pval.size());
+    long long int x, y;
     MathFunc::ext_gcd(s, p - 1, x, y);
 
     for (size_t i = 0; i < pval.size(); ++i) {
@@ -33,12 +33,12 @@ std::vector<int> NSKalgo::publickey(int p, int s) {
 }
 
 // Encryption process
-std::vector<int> NSKalgo::encrypt(const std::vector<int>& mess, const std::vector<int>& pubk, int p) {
-    std::vector<int> cipher(mess.size());
+std::vector<long long int> NSKalgo::encrypt(const std::vector<long long int>& mess, const std::vector<long long int>& pubk, long long int p) {
+    std::vector<long long int> cipher(mess.size());
     
     for (size_t i = 0; i < mess.size(); ++i) {
-        std::vector<int> binaryArray = MathFunc::intToBinArr(mess[i]);  // Get binary representation
-        int c = 1;
+        std::vector<long long int> binaryArray = MathFunc::intToBinArr(mess[i]);  // Get binary representation
+        long long int c = 1;
         for (size_t j = 0; j < pubk.size(); ++j) {
             c = (c * MathFunc::mod_exp(pubk[j], binaryArray[j], p)) % p;  // Multiply with public key
         }
@@ -49,15 +49,15 @@ std::vector<int> NSKalgo::encrypt(const std::vector<int>& mess, const std::vecto
 }
 
 // Decryption process
-std::vector<int> NSKalgo::decrypt(const std::vector<int>& cipher, int p, int s) {
-    std::vector<int> decipher(cipher.size());
+std::vector<long long int> NSKalgo::decrypt(const std::vector<long long int>& cipher, long long int p, long long int s) {
+    std::vector<long long int> decipher(cipher.size());
 
     for (size_t i = 0; i < cipher.size(); ++i) {
-        int val = 0;
+        long long int val = 0;
         for (size_t j = 0; j < pval.size(); ++j) {
-            int term1 = MathFunc::mod_exp(2, j, p);
-            int term2 = MathFunc::invmod(pval[j] - 1, p);
-            int term3 = MathFunc::itemgcd(pval[j], MathFunc::mod_exp(cipher[i], s, p)) - 1;
+            long long int term1 = MathFunc::mod_exp(2, j, p);
+            long long int term2 = MathFunc::invmod(pval[j] - 1, p);
+            long long int term3 = MathFunc::itemgcd(pval[j], MathFunc::mod_exp(cipher[i], s, p)) - 1;
             val = (val + ((term1 * term2) % p) * term3) % p;
         }
         decipher[i] = val;
